% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.character}
\alias{as.character}
\alias{as.character,UUID-method}
\alias{as.character,UUIDList-method}
\title{Coerce UUID(s) to character string(s)}
\usage{
\S4method{as.character}{UUID}(x, upper.case = getOption("UUIDuppercase"))

\S4method{as.character}{UUIDList}(x, upper.case = getOption("UUIDuppercase"))
}
\arguments{
\item{x}{A \code{\linkS4class{UUID}} or \code{\linkS4class{UUIDList}} object}

\item{upper.case}{A logical value specifying whether the result should be returned in upper
case (\code{TRUE}), lower case (\code{FALSE}), or the libuuid default
(\code{NA})}
}
\value{
A character vector containing one element for each \code{UUID};
nil \code{UUID} values are coerced to \code{NA} values.
}
\description{
These methods coerce \code{\linkS4class{UUID}} or
\code{\linkS4class{UUIDList}} objects to character vectors.
}
\details{
\describe{
  \item{\code{UUID}}{
    As S4 method is used, rather than an S3 method \code{as.character.UUID},
    for consistency with the S4 method for \code{as.character} and class
    \code{UUIDList}.
  }
  \item{\code{UUIDList}}{
    Defining an S3 method function \code{as.character.UUIDList} does not work
    because the S4 method \code{\link[S4Vectors]{as.character,Vector-method}}
    is called instead, as \code{UUIDList} inherits
    \code{\linkS4class{Vector}}, and S4 methods are called before S3 methods.
    That S4 method passes its input to \code{\link{as.vector}}, which in turn
    terminates with an error when passed a \code{UUIDList} object, as an
    \code{as.vector} method does not exist for class \code{UUIDList}.
    As there is no way to coerce a \code{UUIDList} object to a vector, an S4
    method for \code{as.character} is defined instead.
  }
}
These methods are also called by the S4 method
\code{\link{coerce,ANY,character-method}}.
}
\author{
Adam C. Gower \email{agower@bu.edu},
based on libuuid by Theodore Ts'o.
}
