% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.raw}
\alias{as.raw}
\alias{as.raw.UUID}
\title{Coerce UUID to raw vector}
\usage{
\method{as.raw}{UUID}(x)
}
\arguments{
\item{x}{A \code{\linkS4class{UUID}} object}
}
\value{
A raw vector of length 16.
}
\description{
This S3 method coerces a \code{\linkS4class{UUID}} object to a raw vector.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
