% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as}
\alias{as}
\alias{coerce,UUID,raw-method}
\title{Coercion methods for UUIDs}
\arguments{
\item{from}{A \code{character} or \code{UUID} object}
}
\value{
An R object of the target class.
}
\description{
These methods coerce \code{character} objects to \code{\linkS4class{UUID}} or
\code{\linkS4class{UUIDList}} objects, or \code{UUID} objects to
\code{UUIDList} objects.
}
\author{
Adam C. Gower \email{agower@bu.edu},
based on libuuid by Theodore Ts'o.
}
