% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{Comparison}
\alias{Comparison}
\alias{==,UUID,UUID-method}
\alias{!=,UUID,UUID-method}
\alias{<=,UUID,UUID-method}
\alias{>=,UUID,UUID-method}
\alias{<,UUID,UUID-method}
\alias{>,UUID,UUID-method}
\title{Relational operators for UUIDs}
\usage{
\S4method{==}{UUID,UUID}(e1, e2)

\S4method{!=}{UUID,UUID}(e1, e2)

\S4method{<=}{UUID,UUID}(e1, e2)

\S4method{>=}{UUID,UUID}(e1, e2)

\S4method{<}{UUID,UUID}(e1, e2)

\S4method{>}{UUID,UUID}(e1, e2)
}
\arguments{
\item{e1}{A \code{\linkS4class{UUID}} object}

\item{e2}{A \code{\linkS4class{UUID}} object}
}
\value{
A logical vector of length 1.
}
\description{
These methods for binary operators use \code{\link{pcompare}} to perform
tests between pairs of \code{\linkS4class{UUID}} objects. They are identical
to the corresponding methods for \code{\linkS4class{Vector}} objects in
package \pkg{S4Vectors}.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
