% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{duplicated}
\alias{duplicated}
\alias{duplicated,UUIDList-method}
\title{Determine duplicated UUIDs}
\usage{
\S4method{duplicated}{UUIDList}(x, incomparables = FALSE, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{UUIDList}} object}

\item{incomparables}{A \code{\linkS4class{UUID}} or a \code{UUIDList} containing values that
should be excluded from comparison, or the logical value \code{FALSE},
meaning that all values can be compared.}

\item{\dots}{Additional arguments passed to
\code{\link[S4Vectors]{duplicated,Vector-method}} and to \code{\link{match}}}
}
\value{
A logical vector of the same length as \code{x} indicating whether each
element in \code{x} is duplicated.
}
\description{
This method provides the ability to identify duplicate elements in a
\code{\linkS4class{UUIDList}} object.
}
\details{
This method is defined only to ensure that the \code{Vector} method for
\code{duplicate} is used, rather than the \code{List} method, which
supersedes it for \code{UUIDList} objects.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
