% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{isNil}
\alias{isNil}
\alias{isNil,UUID-method}
\alias{isNil,UUIDList-method}
\title{Test for nil UUIDs}
\usage{
isNil(x)

\S4method{isNil}{UUID}(x)

\S4method{isNil}{UUIDList}(x)
}
\arguments{
\item{x}{A \code{\linkS4class{UUID}} or \code{\linkS4class{UUIDList}} object}
}
\value{
A logical vector of the same length as the input; if each \code{UUID} is nil,
\code{TRUE}; otherwise, \code{FALSE}.
}
\description{
These methods test whether a \code{\linkS4class{UUID}} object (or each
element of a \code{\linkS4class{UUIDList}} object) is a "nil UUID",
i.e., consists solely of zero bits.
It is analogous to the \code{\link{is.na}} function for atomic vectors.
}
\references{
https://www.ietf.org/rfc/rfc4122.txt
}
\author{
Adam C. Gower \email{agower@bu.edu},
based on libuuid by Theodore Ts'o.
}
