% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{order}
\alias{order}
\alias{order,UUIDList-method}
\title{UUIDList ordering permutation}
\usage{
\S4method{order}{UUIDList}(
  ...,
  na.last = TRUE,
  decreasing = FALSE,
  method = c("shell", "quick", "radix")
)
}
\arguments{
\item{\dots}{A sequence of \code{\linkS4class{UUIDList}} objects, all of the same length.}

\item{na.last}{A logical value for controlling the treatment of \code{NA}s.
If \code{TRUE}, missing values in the data are put last;
if \code{FALSE}, they are put first; if \code{NA}, they are removed.}

\item{decreasing}{A logical value specifying whether the sort order should be increasing or
decreasing.}

\item{method}{A character string specifying the method to be used: partial matches are
allowed.  For details of specific methods, see the help for
\code{\link{sort}}.}
}
\value{
An integer vector of the same length as the inputs.
}
\description{
This method for \code{\linkS4class{UUIDList}} objects returns a permutation
that rearranges its first argument into ascending or descending order,
breaking ties by further arguments.
}
\seealso{
This method enables use of the \code{\linkS4class{Vector}} methods for
\code{\link{sort}} and (together with \code{\link{as.character}})
\code{\link{table}} (see Note in \code{\link{pcompare}}).
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
