% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pcompare}
\alias{pcompare}
\alias{pcompare,UUID,UUID-method}
\alias{pcompare,UUIDList,UUID-method}
\alias{pcompare,UUID,UUIDList-method}
\alias{pcompare,UUIDList,UUIDList-method}
\title{Compare UUIDs}
\usage{
\S4method{pcompare}{UUID,UUID}(x, y)

\S4method{pcompare}{UUIDList,UUID}(x, y)

\S4method{pcompare}{UUID,UUIDList}(x, y)

\S4method{pcompare}{UUIDList,UUIDList}(x, y)
}
\arguments{
\item{x}{A \code{\linkS4class{UUID}} or \code{\linkS4class{UUIDList}} object}

\item{y}{A \code{\linkS4class{UUID}} or \code{\linkS4class{UUIDList}} object}
}
\value{
An integer vector containing the result of the element-by-element comparison:
\describe{
  \item{\code{-1}}{
    The \code{UUID} element of \code{x} is less than the corresponding
    \code{UUID} element of \code{y}
  }
  \item{\code{0}}{
    The \code{UUID} element of \code{x} is equal to the corresponding
    \code{UUID} element of \code{y}
  }
  \item{\code{1}}{
    The \code{UUID} element of \code{x} is greater than the corresponding
    \code{UUID} element of \code{y}
  }
  \item{\code{NA}}{
    The \code{UUID} element of \code{x} or \code{y} is nil (this is for
    consistency with atomic comparisons with \code{NA} values, which also
    return \code{NA})
  }
}
The elements of shorter vectors are recycled as necessary.
}
\description{
These methods provide the ability to compare \code{\linkS4class{UUID}} and/or
\code{\linkS4class{UUIDList}} objects.
}
\author{
Adam C. Gower \email{agower@bu.edu},
based on libuuid by Theodore Ts'o.
}
