% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{show}
\alias{show}
\alias{show,UUID-method}
\title{Describe a UUID}
\usage{
\S4method{show}{UUID}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{UUID}} object}
}
\description{
This method shows the contents of a \code{\linkS4class{UUID}}, including its
variant and version numbers.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
