% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{UUID-class}
\alias{UUID-class}
\alias{UUID}
\title{Class to contain a Universally Unique Identifier (UUID)}
\usage{
UUID(data)
}
\arguments{
\item{data}{A raw vector of length 16.}
}
\description{
This S4 class is a container for a raw vector of length 16,
specifying a 128-bit UUID.
}
\section{Slots}{

\describe{
\item{\code{data}}{A raw vector of length 16.
Defaults to a vector of zeroes, i.e., a "nil UUID".}
}}

\references{
https://www.ietf.org/rfc/rfc4122.txt
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
