% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UUIDtime.R
\name{UUIDtime}
\alias{UUIDtime}
\title{Get the time of a UUID}
\usage{
UUIDtime(x, tz = "")
}
\arguments{
\item{x}{A \code{\linkS4class{UUID}} object}

\item{tz}{A time zone specification to be used for converting time;
defaults to \code{""}, i.e., the current time zone}
}
\value{
A \code{\link{POSIXct}} object specifying the time at which the UUID was
created, in seconds since January 1, 1970 GMT (the epoch).
}
\description{
This function gets the time at which a \code{UUID} was generated.
}
\references{
https://www.ietf.org/rfc/rfc4122.txt
}
\author{
Adam C. Gower \email{agower@bu.edu},
based on libuuid by Theodore Ts'o.
}
