% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UUIDvariant.R
\name{UUIDvariant}
\alias{UUIDvariant}
\title{Get the variant of a UUID}
\usage{
UUIDvariant(x)
}
\arguments{
\item{x}{A \code{\linkS4class{UUID}} object}
}
\value{
An integer value indicating the variant of the UUID:
\describe{
  \item{\code{0}}{Reserved, NCS backward compatibility}
  \item{\code{1}}{DCE specification}
  \item{\code{2}}{Reserved, Microsoft Corporation backward compatibility}
  \item{\code{3}}{Reserved for future definition}
}
}
\description{
This function gets the bits of a \code{\linkS4class{UUID}} object
specifying its variant (i.e., format).
}
\references{
https://www.ietf.org/rfc/rfc4122.txt
}
\author{
Adam C. Gower \email{agower@bu.edu},
based on libuuid by Theodore Ts'o.
}
