% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UUIDversion.R
\name{UUIDversion}
\alias{UUIDversion}
\title{Get the version of a UUID}
\usage{
UUIDversion(x)
}
\arguments{
\item{x}{A \code{\linkS4class{UUID}} object}
}
\value{
An integer value indicating the version of the UUID:
\enumerate{
  \item{Time-based UUIDs}
  \item{DCE security version}
  \item{Namespace-based version using MD5 hashing}
  \item{Randomly generated}
  \item{Namespace-based version using SHA1 hashing}
}
}
\description{
This function gets the bits of a \code{\linkS4class{UUID}} object
specifying its version (i.e., how it was generated).
}
\references{
https://www.ietf.org/rfc/rfc4122.txt
}
\author{
Adam C. Gower \email{agower@bu.edu},
based on libuuid by Theodore Ts'o.
}
